requirements.txt
asgiref~=3.3.4
channels~=3.0.3
channels-redis==3.2.0
colorama==0.4.4
decorator==4.4.2
Django>=3.0,<4.0
django-cors-headers==3.7.0
django-dbbackup==3.3.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-q==1.3.9
django-rest-auth==0.9.5
djangorestframework==3.12.4
docker==5.0.0
filetype==1.0.7
gunicorn==20.1.0
jsonfield==3.1.0
numpy==1.20.3
opencv-python==4.5.2.52
openpyxl==3.0.7
Pillow==8.2.0
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
requests~=2.25.1
rncryptor==3.3.0
simplejson==3.17.2
uvicorn==0.14.0
xlrd==2.0.1
xlsxwriter==1.4.3
psycopg2-binary>=2.8

Faker==13.3.2
pandas==1.4.1
python-memcached>=1.3

